abstract api log into service
