+++ packages/server-core/src/api/api-log.service.ts
@@ -1,0 +1,36 @@
+import { isDev } from '@etherealengine/common/src/config'
+
+import { Application } from '../../declarations'
+import { elasticOnlyLogger } from '../ServerLogger'
+
+declare module '@etherealengine/common/declarations' {
+  interface ServiceTypes {
+    '/api/log': any
+  }
+}
+
+// Receive client-side log events (only active when APP_ENV != 'development')
+export default (app: Application): void => {
+  app.use(
+    '/api/log',
+    {
+      create: () => {
+        return
+      }
+    },
+    {
+      koa: {
+        before: [
+          async (ctx: any, next) => {
+            console.log(ctx)
+            const { msg, ...mergeObject } = ctx.body
+            if (!isDev) elasticOnlyLogger.info({ user: ctx.params?.user, ...mergeObject }, msg)
+            await next()
+            ctx.status = 204
+            return
+          }
+        ]
+      }
+    }
+  )
+}
