+++ packages/server-core/src/api/services.ts
@@ -1,0 +1,3 @@
+import apiLogService from './api-log.service'
+
+export default [apiLogService]
